+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ Changes in 2.2.0
 
 * Added known format as ``KNOWN_FORMATS``.
 
+* Added ability to force write strings of numbers without quotes, through the
+  ``quote_numbers`` argument.
+
+* Added ability to not evaluate numbers without quotes upon read, through the
+  ``eval_cells11 argument.
+
 
 Changes in 2.1.0
 ================